refactor(user-page): move fork switch to user page component
